pyproject.toml
[project]
name = "fenix-prospector"
version = "0.1.0"
description = "Lead prospecting tool from Spanish public registries (BOE, BORME)"
requires-python = ">=3.11"

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "jinja2>=3.1.4",
    "httpx>=0.28.0",
    "sqlalchemy>=2.0.36",
    "aiosqlite>=0.20.0",
    "alembic>=1.14.0",
    "pdfminer.six>=20231228",
    "lxml>=5.3.0",
    "openpyxl>=3.1.5",
    "apscheduler>=3.10.4",
    "pydantic-settings>=2.7.0",
    "python-multipart>=0.0.18",
    "unidecode>=1.3.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "ruff>=0.8.0",
]
